refactor(api): reorganize API routes for clarity (#99)
