+++ apps/api/src/types/env.ts
@@ -1,0 +1,5 @@
+export type Env = {
+  DATABASE_URL: string;
+  REDIS_URL: string;
+  REDIS_TOKEN: string;
+};
